Revert "[VCDA-3430] Restore 1.0.0 rde types and update 1.1.0 types as per the new schema (#79)"
This reverts commit 9fbec60932a204d12d3093b12f4e664293024421.
